pyproject.toml
[project]
name = "docfusion"
version = "1.0.0"
description = "DocFusion: Operation Intelligent Documents - rihal CodeStacker 2026 ML Challenge"
requires-python = ">=3.13"
authors = [{ name = "0xabdulraheem" }]

dependencies = [
    "easyocr>=1.7.2",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "pillow>=11.0.0",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.6.0",
    "lightgbm>=4.5.0",
    "scipy>=1.14.0",
    "python-dateutil>=2.9.0",
    "rapidfuzz>=3.10.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "opencv-python-headless>=4.10.0",
    "joblib>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.1.0",
    "ipykernel>=6.29.0",
    "pytest>=8.3.0",
]
datasets = [
    "datasets>=3.1.0",
    "kagglehub>=0.3.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
